fix: RemainingTime::from_ms() truncation at u8 boundary
Cast ms/1000 to u32 before divmod instead of u8.
Previously 300s (5min) truncated to 44, 600s (10min) to 88,
causing the display to show wrong countdown from the start.

+++ src/domain/time.rs
@@ -8,10 +8,10 @@ pub struct RemainingTime {
 impl RemainingTime {
     /// Build from remaining milliseconds.
     pub const fn from_ms(ms: u64) -> Self {
-        let total_secs = (ms / 1_000) as u8;
+        let total_secs = (ms / 1_000) as u32;
         Self {
-            minutes: total_secs / 60,
-            seconds: total_secs % 60,
+            minutes: (total_secs / 60) as u8,
+            seconds: (total_secs % 60) as u8,
         }
     }
